requirements.txt
click==8.0.1
dominate==2.6.0
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.21.2
Pillow==8.3.2
SQLAlchemy==1.4.25
torch==1.9.1
torchvision==0.10.1
typing-extensions==3.10.0.2
visitor==0.1.3
Werkzeug==2.0.1
zipp==3.6.0
